Revert "closes #17"
