--- SOLID/node_importer.routing.yml
@@ -1,15 +1,0 @@
-SOLID:
-    path: '/SOLID'
-    defaults:
-        _title: 'SOLID - Form'
-        _form: '\Drupal\SOLID\Controller\Form'
-    requirements:
-        _role: 'administrator'
-
-SOLID.progress:
-    path: '/SOLID/progress'
-    defaults:
-        _title: 'SOLID - Progress'
-        _controller: '\Drupal\SOLID\Controller\Progress::content'
-    requirements:
-        _role: 'administrator'
