+++ node_importer/node_importer.links.menu.yml
@@ -1,6 +1,6 @@
-SOLID:
+node_importer:
   title: Node Importer
   description: Import JSON/OWL files to automatically create nodes and vocabularies.
-  route_name: SOLID
-  base_route: SOLID
+  route_name: node_importer
+  base_route: node_importer
   parent: system.admin_config_content
