+++ node_importer/node_importer.routing.yml
@@ -1,0 +1,15 @@
+node_importer:
+    path: '/node_importer'
+    defaults:
+        _title: 'NodeImporter - Form'
+        _form: '\Drupal\node_importer\Controller\Form'
+    requirements:
+        _role: 'administrator'
+
+node_importer.progress:
+    path: '/node_importer/progress'
+    defaults:
+        _title: 'NodeImporter - Progress'
+        _controller: '\Drupal\node_importer\Controller\Progress::content'
+    requirements:
+        _role: 'administrator'
