+++ node_importer/src/Script/delete.php
